+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.1.0] - 15/11/2025
+
+### Added
+- Added function to the delete products button
+
 ## [0.1.0] - 14/11/2025
 
 ### Fixed
@@ -14,11 +19,14 @@
 - Added Workshop Card atom.
 - Added screen for view Workshop 
 - Added function for searching workshops by name.
+- Added US to Register Product Batches
 
 ### Fixed
 - ArcaApi get all external collabs
 - Search users
+- Added presentation, data and domain for RegisterSupplyBatch
 - Fix snackbar for register new workshops
+- Fix error of implementation in Input atoms
 
 ## [0.1.0] - 12/11/2025